fix(ci-cd): improve git-cliff installation and tag workflow
- Pin git-cliff to v2.6.0 and add SHA-512 checksum verification for fallback
- Move tag creation to after changelog commit to ensure correct SHA
- Push only the specific version tag instead of all tags
- Centralize APP_VERSION lookup in shared getAppVersion utility
- Update validateEnvironment and healthController to use shared helper
